module integration

+++ backend.py
@@ -10,7 +10,8 @@ import subprocess
 from subprocess import run
 import time
 from absl import logging
-
+import os.path
+from os import path
 
 import tensorflow as tf
 import tensorflow_hub as hub
@@ -26,42 +27,12 @@ import seaborn as sns
   
 # module_url = "D:/sem8/module5"
 
-module_url = "/users/anagh/Desktop/module5"
-embed = hub.KerasLayer(module_url)
-print("module loaded")
+# module_url = "/users/anagh/Desktop/module5"
+# embed = hub.KerasLayer(module_url)
+# print("module loaded")
 
 
-logging.set_verbosity(logging.ERROR)
-
-def plot_similarity(labels, features, rotation):
-  corr = np.inner(features, features)
-#   sns.set(font_scale=1.2)
-#   g = sns.heatmap(
-#       corr,
-#       xticklabels=labels,
-#       yticklabels=labels,
-#       vmin=0,
-#       vmax=1,
-#       cmap="YlOrRd")
-  # g.set_xticklabels(labels, rotation=rotation)
-  # g.set_title("Semantic Textual Similarity")
-  return corr[0][1]
-
-
-def run_and_plot(session_, input_tensor_, messages_, encoding_tensor):
-  message_embeddings_ = session_.run(
-      encoding_tensor, feed_dict={input_tensor_: messages_})
-  # return plot_similarity(messages_, message_embeddings_, 90)
-
-def find_sim(messages):
-    similarity_input_placeholder = tf.placeholder(tf.string, shape=(None))
-    similarity_message_encodings = embed(similarity_input_placeholder)
-    with tf.Session() as session:
-        session.run(tf.global_variables_initializer())
-        session.run(tf.tables_initializer())
-        return run_and_plot(session, similarity_input_placeholder, messages,
-                    similarity_message_encodings)
-
+# logging.set_verbosity(logging.ERROR)
 
 
 app = Flask(__name__) 
@@ -761,6 +732,18 @@ def getStudents():
         students = []
         try:
             classcode = request.get_data()
+            teacher = session['username']
+            path = os.path.abspath(os.getcwd()) + '/automated_correction_module/' + teacher + '/'
+            if(os.path.exists(path)):
+                print("Directory exists")
+                shutil.rmtree(path)
+                os.makedirs(path)
+
+            else:
+                print("directory does not exist creating directory")
+                os.makedirs(path)
+
+
             mycon = mysql.connector.connect( host="localhost", user="root", passwd="123456789", db="digischool" )
             cursor = mycon.cursor()
             query = "select * from student where classcode=%s"
@@ -791,18 +774,33 @@ def upload():
     if request.method == 'POST':  
         f = request.files['file']  
         f.save(f.filename)
-        print(f.filename)
+        # print(f.filename)
+        teacher = session['username']
         names = request.form['name']
         classcode = request.form['classCode']
         subject = request.form['subject']
         print(names,classcode,subject)
-        print("Name is : " + str(names))
-        print('Classcode ' + str(classcode))
-        print('subject : ' + str(subject))
-        print(f.filename)
+        # print("Name is : " + str(names))
+        # print('Classcode ' + str(classcode))
+        # print('subject : ' + str(subject))
+        # print(f.filename)
         src = f.filename
-        dest = 'automated_correction_module/input/'+str(names)+'_' + str(classcode) + '_' + str(subject) + '.png'
-        shutil.move(src, dest)
+
+        path = os.path.abspath(os.getcwd()) + '/automated_correction_module/' + teacher + '/'
+        print(path)
+        if(os.path.exists(path)):
+            
+            path_files = path + 'input/'
+            if(os.path.exists(path_files)):
+                dest = path_files +str(names)+'_' + str(classcode) + '_' + str(subject) + '.png'
+                shutil.move(src, dest)
+            else:
+                os.makedirs(path_files)
+                dest = path_files +str(names)+'_' + str(classcode) + '_' + str(subject) + '.png'
+                shutil.move(src, dest)
+
+        
+
     data = {}
     data['msg'] = "done"
     print(data)
@@ -813,9 +811,12 @@ def correct():
     if(request.method == 'POST'):
         data = {}
         print("In /correct")
-        path = 'automated_correction_module/input/'
+        teacher = session['username']
+
+        path = 'automated_correction_module/' + teacher + '/input/'
         files = os.listdir(path)
         print(files)
+        recognition_files = []
         for file in files:
             # img_path = 'input/' + file
             if('.png' in file or '.jpg' in file):
@@ -829,21 +830,26 @@ def correct():
                 pipe = subprocess.check_call(["python", "Prediction.py", file] , cwd = initial_directory + '/automated_correction_module/classification/')
                 time.sleep(5)
                 print("/Correct prediction over")
-                f = open("automated_correction_module/recognized.txt", "r")
-                sent1 = f.read()
-                print("Recognized" + str(sent1))
 
+                recognition_files.append(file.split('.png')[0] + '.txt')
                 
-                f2 = open("automated_correction_module/answer_key.txt", "r")
-                sent2 = f2.read()
-                print("Answer Key is : " + str(sent2))
+                
 
-      
-                messages = [sent1,sent2]
-                print(messages)
-                # print("finding similarity\n")
-                # print(find_sim(messages))
+        string = "["
+        for i in range(len(recognition_files)):
+            if(i == len(recognition_files) - 1):
+                string += recognition_files[i] + "]"
+            else:
+                string += recognition_files[i] + ","
 
+        path = os.path.abspath(os.getcwd()) + '/automated_correction_module/' + teacher + '/'
+        files = os.listdir(path)
+        for i in range(len(files)):
+            if '.txt' in files[i]:
+                answer_file = files[i]
+
+        pipe1 = subprocess.check_call(["python", "semantic.py" , string, path + answer_file], cwd= initial_directory + '/automated_correction_module/')
+        time.sleep(5)
         return data
     
 
@@ -1088,15 +1094,19 @@ def uploadAnswerKey():
         f = request.files['file']  
         f.save(f.filename)
         print(f.filename)
-        
+        teacher = session['username']
         classcode = request.form['classCode']
         subject = request.form['subject']
         print('Classcode ' + str(classcode))
         print('subject : ' + str(subject))
         # print(f.filename)
         src = f.filename
-        dest = 'automated_correction_module/'+ str(classcode) + '_' + str(subject) + '.txt'
-        shutil.move(src, dest)
+        path = os.path.abspath(os.getcwd()) + '/automated_correction_module/' + teacher + '/'
+        print(path)
+        if(os.path.exists(path)):
+            dest = path + str(classcode) + '_' + str(subject) + '.txt'
+            shutil.move(src, dest)
+
     data = {}
     data['msg'] = "done"
     print(data)
